# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AegisPulse Clinical Deterioration Radar API
  version: 1.0.0
  description: |
    Production-oriented backend API for the AegisPulse Continuous Patient Deterioration Radar.
    Provides deterministic clinical surveillance, early warning scoring (Subbe MEWS, Singer qSOFA),
    explainability reasoning, unified patient timeline analytics, and ward simulation controls.
servers:
  - url: http://localhost:3000
    description: Local Development / Test Server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/wards/{wardId}/radar:
    get:
      tags: [Wards & Beds, Attention Priorities]
      summary: Ward-wide patient deterioration priority radar
      parameters:
        - name: wardId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ranked list of patients by APS score
  /api/v1/beds:
    get:
      tags: [Wards & Beds]
      summary: List beds
      parameters:
        - name: wardId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of beds
  /api/v1/beds/{bedId}:
    get:
      tags: [Wards & Beds]
      summary: Get bed details and current patient
      parameters:
        - name: bedId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bed details
...
  /api/v1/patients/{patientId}/observations:
    get:
      tags: [Observations]
      summary: Query patient observations
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
        - name: since
          in: query
          schema:
            type: integer
        - name: until
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Observations list
    post:
      tags: [Observations]
      summary: Ingest validated bedside observation
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestObservationRequest'
      responses:
        '21':
          description: Ingestion accepted and indexed
        '400':
          description: Validation error or out of physiological limits
        '429':
          description: Rate limit exceeded
  /api/v1/patients/{patientId}/labs:
    get:
      tags: [Labs]
      summary: Query laboratory results
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Labs list
    post:
      tags: [Labs]
      summary: Ingest laboratory result
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestLabRequest'
      responses:
        '201':
          description: Lab result recorded
...
  /api/v1/patients/{patientId}/attention-priority:
    get:
      tags: [Attention Priorities]
      summary: Real-time Attention Priority Score evaluation and explainability
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: APS score, category, component breakdown, and curated 3-5 reasons
  /api/v1/patients/{patientId}/acknowledgements:
    get:
      tags: [Acknowledgements]
      summary: List alert acknowledgements
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Acknowledgements list
    post:
      tags: [Acknowledgements]
      summary: Acknowledge alert or priority escalation (Staff role required)
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                alertId:
                  type: string
                reason:
                  type: string
      responses:
        '201':
          description: Acknowledged
  /api/v1/patients/{patientId}/actions:
    get:
      tags: [Clinical Actions]
      summary: List clinical actions
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Actions list
    post:
      tags: [Clinical Actions]
      summary: Propose clinical action
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [actionType, title, rationale]
              properties:
                actionType:
                  type: string
                title:
                  type: string
                rationale:
                  type: string
      responses:
        '201':
          description: Action proposed
  /api/v1/patients/{patientId}/actions/{actionId}:
    patch:
      tags: [Clinical Actions]
      summary: Update clinical action status (Mark COMPLETED with notes)
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
        - name: actionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [RECOMMENDED, ACKNOWLEDGED, IN_PROGRESS, COMPLETED, DISMISSED]
                outcomeNotes:
                  type: string
      responses:
        '200':
          description: Action updated
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    IngestObservationRequest:
      type: object
      properties:
        source:
          type: string
        confidence:
          type: number
        qualityState:
          type: string
        heartRate:
          type: number
        respiratoryRate:
          type: number
        systolicBP:
          type: number
        diastolicBP:
          type: number
        temperature:
          type: number
        spo2:
          type: number
        shockIndex:
          type: number
        notes:
          type: string
    IngestLabRequest:
      type: object
      required: [testCode, testName, value, unit, referenceRange]
      properties:
        testCode:
          type: string
        testName:
          type: string
        value:
          type: number
        unit:
          type: string
        referenceRange:
          type: object
          properties:
            low:
              type: number
            high:
              type: number
        isCritical:
          type: boolean
...